requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
aiohappyeyeballs=2.4.0=py312hecd8cb5_0
aiohttp=3.10.5=py312h46256e1_0
aiosignal=1.2.0=pyhd3eb1b0_0
arrow-cpp=16.1.0=hc0a701b_0
attrs=23.1.0=py312hecd8cb5_0
aws-c-auth=0.6.19=h6c40b1e_0
aws-c-cal=0.5.20=h3333b6a_0
aws-c-common=0.8.5=h6c40b1e_0
aws-c-compression=0.2.16=h6c40b1e_0
aws-c-event-stream=0.2.15=hcec6c5f_0
aws-c-http=0.6.25=h6c40b1e_0
aws-c-io=0.13.10=h6c40b1e_0
aws-c-mqtt=0.7.13=h6c40b1e_0
aws-c-s3=0.1.51=h6c40b1e_0
aws-c-sdkutils=0.1.6=h6c40b1e_0
aws-checksums=0.1.13=h6c40b1e_0
aws-crt-cpp=0.18.16=hcec6c5f_0
aws-sdk-cpp=1.10.55=h61975a4_0
blas=1.0=openblas
blinker=1.6.2=py312hecd8cb5_0
boost-cpp=1.82.0=ha357a0b_2
bottleneck=1.3.7=py312h32608ca_0
brotli-python=1.0.9=py312hcec6c5f_8
bzip2=1.0.8=h6c40b1e_6
c-ares=1.19.1=h6c40b1e_0
ca-certificates=2024.7.2=hecd8cb5_0
cairo=1.16.0=h3ce6f7e_5
certifi=2024.8.30=py312hecd8cb5_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
click=8.1.7=py312hecd8cb5_0
cyrus-sasl=2.1.28=h3973f50_1
datasets=2.19.1=py312hecd8cb5_0
diffusers=0.18.2=py312h8e4b320_0
diffusers-base=0.18.2=py312h8e4b320_0
diffusers-torch=0.18.2=py312h8e4b320_0
dill=0.3.8=py312hecd8cb5_0
eigen=3.4.0=ha357a0b_0
expat=2.6.3=h6d0c2b6_0
ffmpeg=4.2.2=h97e5cf8_0
filelock=3.13.1=py312hecd8cb5_0
flask=3.0.3=py312hecd8cb5_0
fontconfig=2.14.1=h269ac6a_3
freetype=2.12.1=hd8bbffd_0
frozenlist=1.4.0=py312h6c40b1e_0
fsspec=2024.3.1=py312hecd8cb5_0
gettext=0.21.0=h4e8c18a_2
gflags=2.2.2=hcec6c5f_1
glib=2.78.4=hcec6c5f_0
glib-tools=2.78.4=hcec6c5f_0
glog=0.5.0=hcec6c5f_1
gmp=6.2.1=he9d5cce_3
gnutls=3.6.15=hed9c0bf_0
graphite2=1.3.14=he9d5cce_1
gst-plugins-base=1.14.1=hcec6c5f_1
gstreamer=1.14.1=h6c40b1e_1
harfbuzz=4.3.0=hffc734d_2
hdf5=1.12.1=ha01d115_3
huggingface_accelerate=0.33.0=py312hecd8cb5_0
huggingface_hub=0.24.6=py312hecd8cb5_0
icu=73.1=hcec6c5f_0
idna=3.7=py312hecd8cb5_0
importlib-metadata=7.0.1=py312hecd8cb5_0
itsdangerous=2.2.0=py312hecd8cb5_0
jinja2=3.1.4=py312hecd8cb5_0
jpeg=9e=h46256e1_3
krb5=1.20.1=h428f121_1
lame=3.100=h1de35cc_0
lcms2=2.12=hf1fd2bf_0
lerc=3.0=he9d5cce_0
libabseil=20240116.2=cxx17_h548131f_0
libboost=1.82.0=hf53b9f2_2
libbrotlicommon=1.0.9=h6c40b1e_8
libbrotlidec=1.0.9=h6c40b1e_8
libbrotlienc=1.0.9=h6c40b1e_8
libclang=14.0.6=default_hd95374b_1
libclang13=14.0.6=default_habbcc1a_1
libcurl=8.9.1=h3a17b82_0
libcxx=14.0.6=h9765a3e_0
libdeflate=1.17=hb664fd8_1
libedit=3.1.20230828=h6c40b1e_0
libev=4.33=h9ed2024_1
libevent=2.1.12=h04015c4_1
libffi=3.4.4=hecd8cb5_1
libgfortran=5.0.0=11_3_0_hecd8cb5_28
libgfortran5=11.3.0=h9dfd629_28
libglib=2.78.4=h19e1a8f_0
libgrpc=1.62.2=hf2926fd_0
libiconv=1.16=h6c40b1e_3
libidn2=2.3.4=h6c40b1e_0
libllvm14=14.0.6=h26321d7_4
libnghttp2=1.57.0=h9beae6a_0
libopenblas=0.3.21=h54e7dc3_0
libopus=1.3.1=h1de35cc_0
libpng=1.6.39=h6c40b1e_0
libpq=12.17=h04015c4_0
libprotobuf=4.25.3=h34eed0b_0
libssh2=1.11.0=hf20ceda_0
libtasn1=4.19.0=h6c40b1e_0
libthrift=0.15.0=h70b4b81_2
libtiff=4.5.1=hcec6c5f_0
libunistring=0.9.10=h9ed2024_0
libvpx=1.7.0=h378b8a2_0
libwebp-base=1.3.2=h6c40b1e_0
libxml2=2.13.1=h6070cd6_2
llvm-openmp=14.0.6=h0dcd299_0
lz4-c=1.9.4=hcec6c5f_1
markupsafe=2.1.3=py312h6c40b1e_0
mpmath=1.3.0=py312hecd8cb5_0
multidict=6.0.4=py312h6c40b1e_0
multiprocess=0.70.15=py312hecd8cb5_0
mysql=5.7.24=h1a8d504_2
ncurses=6.4=hcec6c5f_0
nettle=3.7.3=h230ac6f_1
networkx=3.3=py312hecd8cb5_0
numexpr=2.8.7=py312h7d6adbd_0
numpy=1.26.4=py312h7d6adbd_0
numpy-base=1.26.4=py312hbb3573c_0
opencv=4.10.0=py312h1aebfd0_0
openh264=2.1.1=h8346a28_0
openjpeg=2.5.2=hbf2204d_0
openssl=3.0.15=h46256e1_0
orc=2.0.1=h5747287_0
packaging=24.1=py312hecd8cb5_0
pandas=2.2.2=py312h77d3abe_0
pcre2=10.42=h9b97e30_1
pillow=10.4.0=py312h46256e1_0
pip=24.2=py312hecd8cb5_0
pixman=0.40.0=h9ed2024_1
psutil=5.9.0=py312h6c40b1e_0
pyarrow=16.1.0=py312h77d3abe_0
pysocks=1.7.1=py312hecd8cb5_0
python=3.12.5=hcd54a6c_1
python-dateutil=2.9.0post0=py312hecd8cb5_2
python-dotenv=0.21.0=py312hecd8cb5_0
python-tzdata=2023.3=pyhd3eb1b0_0
python-xxhash=2.0.2=py312h6c40b1e_1
pytorch=2.3.0=cpu_py312h781a981_1
pytorch-cpu=2.3.0=hecd8cb5_1
pytz=2024.1=py312hecd8cb5_0
pyyaml=6.0.1=py312h6c40b1e_0
qt-main=5.15.2=hf83fbd5_10
re2=2022.04.01=he9d5cce_0
readline=8.2=hca72f7f_0
regex=2024.9.11=py312h46256e1_0
requests=2.32.3=py312hecd8cb5_0
safetensors=0.4.4=py312hbc49121_0
setuptools=75.1.0=py312hecd8cb5_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.2.1=h6d0c2b6_0
sqlite=3.45.3=h6c40b1e_0
sympy=1.13.2=py312hecd8cb5_0
tk=8.6.14=h4d00af3_0
tokenizers=0.19.1=py312hfa73619_0
tqdm=4.66.5=py312h8e4b320_0
transformers=4.44.1=py312hecd8cb5_0
typing-extensions=4.11.0=py312hecd8cb5_0
typing_extensions=4.11.0=py312hecd8cb5_0
tzdata=2024a=h04d1e81_0
urllib3=2.2.2=py312hecd8cb5_0
utf8proc=2.6.1=h6c40b1e_1
werkzeug=3.0.3=py312hecd8cb5_0
wheel=0.44.0=py312hecd8cb5_0
x264=1!157.20191217=h1de35cc_0
xxhash=0.8.0=h9ed2024_3
xz=5.4.6=h6c40b1e_1
yaml=0.2.5=haf1e3a3_0
yarl=1.11.0=py312h46256e1_0
zipp=3.17.0=py312hecd8cb5_0
zlib=1.2.13=h4b97444_1
zstd=1.5.5=hc035e20_2
